chore: Update libipld to 0.14 in unixfs and minor cleanup

+++ unixfs/src/dir/builder/custom_pb.rs
@@ -79,7 +79,7 @@ impl<'a> MessageWrite for WriteableCid<'a> {
             Version::V0 => hash_len,
             Version::V1 => {
                 let version_len = 1;
-                let codec_len = sizeof_varint(u64::from(self.0.codec()));
+                let codec_len = sizeof_varint(self.0.codec());
                 version_len + codec_len + hash_len
             }
         }
@@ -93,7 +93,7 @@ impl<'a> MessageWrite for WriteableCid<'a> {
                 // directory; at least go-ipfs 0.5 `ipfs files` denies making a cbor link
                 // but happily accepts and does refs over one.
                 w.write_u8(1)?;
-                w.write_varint(u64::from(self.0.codec()))?;
+                w.write_varint(self.0.codec())?;
             }
         }
